requirements.txt
pynacl==1.2.1
discord.py
pip
youtube_dl
